requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
pulp>=2.7.0
jpholiday>=0.1.9
python-dateutil>=2.8.0
numpy>=1.24.0
